update stm32h7 hal

+++ src/hal/stm32h7/Drivers/STM32H7xx_HAL_Driver/Inc/stm32h7xx_hal_rcc.h
@@ -2,8 +2,6 @@
   ******************************************************************************
   * @file    stm32h7xx_hal_rcc.h
   * @author  MCD Application Team
-  * @version V1.2.0
-  * @date   29-December-2017
   * @brief   Header file of RCC HAL module.
   ******************************************************************************
   * @attention
@@ -2854,15 +2852,15 @@ void     HAL_RCC_CCSCallback(void);
   * @{
   */
 
-#define HSE_TIMEOUT_VALUE          ((int32_t)HSE_STARTUP_TIMEOUT)
-#define HSI_TIMEOUT_VALUE          ((int32_t)2)    /* 2 ms */
-#define HSI48_TIMEOUT_VALUE        ((int32_t)2)    /* 2 ms */
-#define CSI_TIMEOUT_VALUE          ((int32_t)2)    /* 2 ms */
-#define LSI_TIMEOUT_VALUE          ((int32_t)2)    /* 2 ms */
-#define PLL_TIMEOUT_VALUE          ((int32_t)2)    /* 2 ms */
-#define CLOCKSWITCH_TIMEOUT_VALUE  ((int32_t)5000) /* 5 s  */
-#define RCC_DBP_TIMEOUT_VALUE      ((int32_t)100)
-#define RCC_LSE_TIMEOUT_VALUE      ((int32_t)LSE_STARTUP_TIMEOUT)
+#define HSE_TIMEOUT_VALUE          HSE_STARTUP_TIMEOUT
+#define HSI_TIMEOUT_VALUE          ((uint32_t)2)    /* 2 ms */
+#define HSI48_TIMEOUT_VALUE        ((uint32_t)2)    /* 2 ms */
+#define CSI_TIMEOUT_VALUE          ((uint32_t)2)    /* 2 ms */
+#define LSI_TIMEOUT_VALUE          ((uint32_t)2)    /* 2 ms */
+#define PLL_TIMEOUT_VALUE          ((uint32_t)2)    /* 2 ms */
+#define CLOCKSWITCH_TIMEOUT_VALUE  ((uint32_t)5000) /* 5 s  */
+#define RCC_DBP_TIMEOUT_VALUE      ((uint32_t)100)
+#define RCC_LSE_TIMEOUT_VALUE      LSE_STARTUP_TIMEOUT
 
 /**
   * @}
